corregido productoTresReales

+++ Producto.java
@@ -34,7 +34,7 @@ public class Producto {
 		 * @return Producto de los tres números
 		 */
  
-		public int productoTresReales(int n1, int n2, int n3) {
+		public double productoTresReales(double n1, double n2, double n3) {
 			return n1*n2*n3;
 		}
